+++ hiredis.c
@@ -312,12 +312,12 @@ static size_t bulklen(size_t len) {
 int redisvFormatCommand(char **target, const char *format, va_list ap) {
     const char *c = format;
     char *cmd = NULL; /* final command */
-    int pos; /* position in final command */
+    size_t pos; /* position in final command */
     sds curarg, newarg; /* current argument */
     int touched = 0; /* was the current argument touched? */
     char **curargv = NULL, **newargv = NULL;
     int argc = 0;
-    int totlen = 0;
+    size_t totlen = 0;
     int error_type = 0; /* 0 = no error; -1 = memory error; -2 = format error */
     int j;
 
@@ -516,7 +516,7 @@ int redisvFormatCommand(char **target, const char *format, va_list ap) {
 
     hi_free(curargv);
     *target = cmd;
-    return totlen;
+    return (int)totlen;
 
 format_err:
     error_type = -2;
@@ -576,7 +576,7 @@ long long redisFormatSdsCommandArgv(sds *target, int argc, const char **argv,
                                     const size_t *argvlen)
 {
     sds cmd, aux;
-    unsigned long long totlen, len;
+    size_t totlen, len;
     int j;
 
     /* Abort on a NULL target */
@@ -616,7 +616,7 @@ long long redisFormatSdsCommandArgv(sds *target, int argc, const char **argv,
     assert(sdslen(cmd)==totlen);
 
     *target = cmd;
-    return totlen;
+    return (long long) totlen; /* api doesn't use ssize_t */
 }
 
 void redisFreeSdsCommand(sds cmd) {
@@ -663,7 +663,7 @@ long long redisFormatCommandArgv(char **target, int argc, const char **argv, con
     cmd[pos] = '\0';
 
     *target = cmd;
-    return totlen;
+    return (long long) totlen;  /* api doesn't use ssize_t */
 }
 
 void redisFreeCommand(char *cmd) {
@@ -941,7 +941,7 @@ redisPushFn *redisSetPushCallback(redisContext *c, redisPushFn *fn) {
  * see if there is a reply available. */
 int redisBufferRead(redisContext *c) {
     char buf[1024*16];
-    int nread;
+    ssize_t nread;
 
     /* Return early when the context has seen an error. */
     if (c->err)
